Extract hook task target app resolution into dedicated step
